requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
black==24.4.2; python_version >= '3.8'
certifi==2024.6.2; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
coverage[toml]==7.5.3; python_version >= '3.8'
dnspython==2.6.1; python_version >= '3.8'
email-validator==2.1.1; python_version >= '3.8'
fastapi[all]==0.111.0; python_version >= '3.8'
fastapi-cli==0.0.4; python_version >= '3.8'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.2.0
jinja2==3.1.4; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
orjson==3.10.3; python_version >= '3.8'
packaging==24.0; python_version >= '3.7'
paho-mqtt==2.1.0; python_version >= '3.7'
pathspec==0.12.1; python_version >= '3.8'
platformdirs==4.2.2; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pydantic==2.7.3; python_version >= '3.8'
pydantic-core==2.18.4; python_version >= '3.8'
pydantic-extra-types==2.8.0
pydantic-settings==2.3.0
pygments==2.18.0; python_version >= '3.8'
pytest==8.2.2; python_version >= '3.8'
pytest-cov==5.0.0; python_version >= '3.8'
pytest-env==1.1.3; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.9; python_version >= '3.8'
pyyaml==6.0.1
rich==13.7.1; python_full_version >= '3.7.0'
ruff==0.9.1; python_version >= '3.7'
shellingham==1.5.4; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.30; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
typer==0.12.3; python_version >= '3.7'
typing-extensions==4.12.1; python_version >= '3.8'
ujson==5.10.0; python_version >= '3.8'
uvicorn[standard]==0.30.1; python_version >= '3.8'
uvloop==0.19.0
watchfiles==0.22.0
websockets==14.1; python_version >= '3.9'
